+++ src/main/java/com/elleined/forumapi/service/react/PostReactionService.java
@@ -91,7 +91,9 @@ public class PostReactionService implements ReactionService<Post, PostReact> {
     }
 
     @Override
-    public void delete(Post post, PostReact postReact) {
+    public void delete(User currentUser, Post post, PostReact postReact) {
+        if (currentUser.notOwned(postReact))
+            throw new NotOwnedException("Cannot delete this post reaction! because you dont owned this reaction!");
         postReactRepository.delete(postReact);
         log.debug("Reaction with id of {} removed successfully with post with id of {}", postReact.getId(), post.getId());
     }
