+++ src/main/java/com/elleined/forumapi/service/react/ReplyReactionService.java
@@ -93,7 +93,9 @@ public class ReplyReactionService implements ReactionService<Reply, ReplyReact> 
     }
 
     @Override
-    public void delete(Reply reply, ReplyReact replyReact) {
+    public void delete(User currentUser, Reply reply, ReplyReact replyReact) {
+        if (currentUser.notOwned(replyReact))
+            throw new NotOwnedException("Cannot delete this reply reaction! because you don't owned this reaction!");
         replyReactRepository.delete(replyReact);
         log.debug("Reaction with id of {} removed successfully with post with id of {}", replyReact.getId(), reply.getId());
     }
